feat: Support config file

+++ src/responses.rs
@@ -32,7 +32,7 @@ pub struct TokenResponse {
     pub access_token: String,
     pub refresh_token: String,
     pub id_token: String,
-    pub expires_in: u32,
+    pub expires_in: u64,
     pub token_type: String
 }
